refactor: :art: code refactor
tried to separate concerns out of one big file into smaller single responsibility files

none
