+++ weather-app-project/index.html
@@ -62,7 +62,7 @@
         >
           <div class="lg:pt-32 lg:pl-10">
             <h1 class="pb-5 text-center">Today's conditions</h1>
-            <h1 id="chosen-location" class="text-4xl pb-2">Johannesburg</h1>
+            <h1 id="chosen-location" class="text-4xl pb-2"></h1>
             <div class="flex justify-center gap-5 items-center">
               <div id="weather-icon-today"></div>
               <div class="flex flex-col">
@@ -73,7 +73,7 @@
           </div>
           <div class="lg:flex lg:justify-center lg:items-center lg:w-full">
             <div
-              class="grid grid-cols-3 gap-2 md:grid-cols-3 text-sm mb-4 lg:justify-center lg:flex-wrap lg:gap-6"
+              class="grid grid-cols-3 gap-2 md:grid-cols-3 text-sm mb-4 lg:justify-center lg:flex-wrap lg:gap-12"
             >
               <div class="flex flex-col items-center gap-3">
                 <svg
@@ -206,108 +206,108 @@
         <div
           class="bg-weekDays rounded-lg p-2 h-20 sm:h-24 md:h-28 lg:h-32 xl:h-72 flex justify-around items-center lg:flex-col"
         >
-          <div id="forecast-two-date">7/11</div>
+          <div id="forecast-2-date"></div>
           <div>
-            <div id="forecast-two-high" class="text-3xl"></div>
+            <div id="forecast-2-high" class="text-3xl"></div>
             <div>high</div>
           </div>
           <div>
-            <div id="forecast-two-low" class="text-3xl"></div>
+            <div id="forecast-2-low" class="text-3xl"></div>
             <div>low</div>
           </div>
           <div class="flex flex-col items-center">
-            <div id="forecast-two-icon"></div>
-            <div id="general-weather-two">sunny</div>
+            <div id="forecast-2-icon"></div>
+            <div id="general-weather-2">sunny</div>
           </div>
         </div>
 
         <div
           class="bg-weekDays rounded-lg p-2 h-20 sm:h-24 md:h-28 lg:h-32 xl:h-72 flex justify-around items-center lg:flex-col"
         >
-          <div id="forecast-three-date"></div>
+          <div id="forecast-3-date"></div>
           <div>
-            <div id="forecast-three-high" class="text-3xl"></div>
+            <div id="forecast-3-high" class="text-3xl"></div>
             <div>high</div>
           </div>
           <div>
-            <div id="forecast-three-low" class="text-3xl"></div>
+            <div id="forecast-3-low" class="text-3xl"></div>
             <div>low</div>
           </div>
           <div class="flex flex-col items-center">
-            <div id="forecast-three-icon"></div>
-            <div id="general-weather-three">sunny</div>
+            <div id="forecast-3-icon"></div>
+            <div id="general-weather-3">sunny</div>
           </div>
         </div>
 
         <div
           class="bg-weekDays rounded-lg p-2 h-20 sm:h-24 md:h-28 lg:h-32 xl:h-72 flex justify-around items-center lg:flex-col"
         >
-          <div id="forecast-four-date"></div>
+          <div id="forecast-4-date"></div>
           <div>
-            <div id="forecast-four-high" class="text-3xl"></div>
+            <div id="forecast-4-high" class="text-3xl"></div>
             <div>high</div>
           </div>
           <div>
-            <div id="forecast-four-low" class="text-3xl"></div>
+            <div id="forecast-4-low" class="text-3xl"></div>
             <div>low</div>
           </div>
           <div class="flex flex-col items-center">
-            <div id="forecast-four-icon"></div>
-            <div id="general-weather-four">sunny</div>
+            <div id="forecast-4-icon"></div>
+            <div id="general-weather-4">sunny</div>
           </div>
         </div>
 
         <div
           class="bg-weekDays rounded-lg p-2 h-20 sm:h-24 md:h-28 lg:h-32 xl:h-72 flex justify-around items-center lg:flex-col"
         >
-          <div id="forecast-five-date"></div>
+          <div id="forecast-5-date"></div>
           <div>
-            <div id="forecast-five-high" class="text-3xl"></div>
+            <div id="forecast-5-high" class="text-3xl"></div>
             <div>high</div>
           </div>
           <div>
-            <div id="forecast-five-low" class="text-3xl"></div>
+            <div id="forecast-5-low" class="text-3xl"></div>
             <div>low</div>
           </div>
           <div class="flex flex-col items-center">
-            <div id="forecast-five-icon"></div>
-            <div id="general-weather-five">sunny</div>
+            <div id="forecast-5-icon"></div>
+            <div id="general-weather-5">sunny</div>
           </div>
         </div>
 
         <div
           class="bg-weekDays rounded-lg p-2 h-20 sm:h-24 md:h-28 lg:h-32 xl:h-72 flex justify-around items-center lg:flex-col"
         >
-          <div id="forecast-six-date"></div>
+          <div id="forecast-6-date"></div>
           <div>
-            <div id="forecast-six-high" class="text-3xl"></div>
+            <div id="forecast-6-high" class="text-3xl"></div>
             <div>high</div>
           </div>
           <div>
-            <div id="forecast-six-low" class="text-3xl"></div>
+            <div id="forecast-6-low" class="text-3xl"></div>
             <div>low</div>
           </div>
           <div class="flex flex-col items-center">
-            <div id="forecast-six-icon"></div>
-            <div id="general-weather-six">sunny</div>
+            <div id="forecast-6-icon"></div>
+            <div id="general-weather-6">sunny</div>
           </div>
         </div>
 
         <div
           class="bg-weekDays rounded-lg p-2 h-20 sm:h-24 md:h-28 lg:h-32 xl:h-72 flex justify-around items-center lg:flex-col"
         >
-          <div id="forecast-seven-date"></div>
+          <div id="forecast-7-date"></div>
           <div>
-            <div id="forecast-seven-high" class="text-3xl"></div>
+            <div id="forecast-7-high" class="text-3xl"></div>
             <div>high</div>
           </div>
           <div>
-            <div id="forecast-seven-low" class="text-3xl"></div>
+            <div id="forecast-7-low" class="text-3xl"></div>
             <div>low</div>
           </div>
           <div class="flex flex-col items-center">
-            <div id="forecast-seven-icon"></div>
-            <div id="general-weather-seven">sunny</div>
+            <div id="forecast-7-icon"></div>
+            <div id="general-weather-7">sunny</div>
           </div>
         </div>
       </div>
